Add tests for the Dynamo-backed IngestVersionManager
