+++ common/src/main/scala/uk/ac/wellcome/platform/archive/common/versioning/dynamo/DynamoIngestVersionManager.scala
@@ -1,0 +1,5 @@
+package uk.ac.wellcome.platform.archive.common.versioning.dynamo
+
+import uk.ac.wellcome.platform.archive.common.versioning.IngestVersionManager
+
+class DynamoIngestVersionManager(val dao: DynamoIngestVersionManagerDao) extends IngestVersionManager
