+++ common/src/test/scala/uk/ac/wellcome/platform/archive/common/versioning/dynamo/DynamoIngestVersionManagerDaoTest.scala
@@ -1,78 +1,17 @@
 package uk.ac.wellcome.platform.archive.common.versioning.dynamo
 
 import com.amazonaws.services.dynamodbv2.model._
-import com.amazonaws.services.dynamodbv2.util.TableUtils.waitUntilActive
 import com.gu.scanamo.{Scanamo, Table => ScanamoTable}
 import uk.ac.wellcome.fixtures.TestWith
 import uk.ac.wellcome.platform.archive.common.bagit.models.ExternalIdentifier
 import uk.ac.wellcome.platform.archive.common.ingests.models.IngestID
 import uk.ac.wellcome.platform.archive.common.versioning.{IngestVersionManagerDao, IngestVersionManagerDaoTestCases, VersionRecord}
 import uk.ac.wellcome.storage.dynamo._
-import uk.ac.wellcome.storage.fixtures.LocalDynamoDb
 import uk.ac.wellcome.storage.fixtures.LocalDynamoDb.Table
 
-import scala.collection.JavaConverters._
 import scala.util.Failure
 
-class DynamoIngestVersionManagerDaoTest extends IngestVersionManagerDaoTestCases[Table] with LocalDynamoDb {
-  override def createTable(table: LocalDynamoDb.Table): LocalDynamoDb.Table = {
-    dynamoDbClient.createTable(
-      new CreateTableRequest()
-        .withTableName(table.name)
-        .withKeySchema(new KeySchemaElement()
-          .withAttributeName("externalIdentifier")
-          .withKeyType(KeyType.HASH))
-        .withKeySchema(new KeySchemaElement()
-          .withAttributeName("version")
-          .withKeyType(KeyType.RANGE))
-        .withAttributeDefinitions(
-          new AttributeDefinition()
-            .withAttributeName("externalIdentifier")
-            .withAttributeType("S"),
-          new AttributeDefinition()
-            .withAttributeName("ingestId")
-            .withAttributeType("S"),
-          new AttributeDefinition()
-            .withAttributeName("version")
-            .withAttributeType("N")
-        )
-        .withGlobalSecondaryIndexes(
-          new GlobalSecondaryIndex()
-            .withIndexName(table.index)
-            .withProjection(
-              new Projection()
-                .withProjectionType(ProjectionType.INCLUDE)
-                .withNonKeyAttributes(
-                  List(
-                    "externalIdentifier",
-                    "ingestId",
-                    "version",
-                    "ingestDate").asJava)
-            )
-            .withKeySchema(
-              new KeySchemaElement()
-                .withAttributeName("ingestId")
-                .withKeyType(KeyType.HASH)
-            )
-            .withProvisionedThroughput(new ProvisionedThroughput()
-              .withReadCapacityUnits(1L)
-              .withWriteCapacityUnits(1L))
-        )
-        .withProvisionedThroughput(new ProvisionedThroughput()
-          .withReadCapacityUnits(1L)
-          .withWriteCapacityUnits(1L))
-    )
-    eventually {
-      waitUntilActive(dynamoDbClient, table.name)
-    }
-    table
-  }
-
-  override def withContext[R](testWith: TestWith[Table, R]): R =
-    withLocalDynamoDbTable { table =>
-      testWith(table)
-    }
-
+class DynamoIngestVersionManagerDaoTest extends IngestVersionManagerDaoTestCases[Table] with IngestVersionManagerTable {
   override def withDao[R](initialRecords: Seq[VersionRecord])(testWith: TestWith[IngestVersionManagerDao, R])(implicit table: Table): R = {
     Scanamo.exec(dynamoDbClient)(ScanamoTable[VersionRecord](table.name).putAll(initialRecords.toSet))
